backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
zhipuai==2.0.1
playwright==1.41.0
markdownify==0.11.6
PyMuPDF==1.23.8
Pillow==10.2.0
requests==2.31.0
pydantic==2.5.3
